+++ ggml/src/ggml-cpu/CMakeLists.txt
@@ -10,8 +10,8 @@ function(ggml_add_cpu_backend_variant_impl tag_name)
     list (APPEND GGML_CPU_SOURCES
         ggml-cpu/ggml-cpu.c
         ggml-cpu/ggml-cpu.cpp
-        ggml-cpu/ggml-cpu-aarch64.cpp
-        ggml-cpu/ggml-cpu-aarch64.h
+        ggml-cpu/repack.cpp
+        ggml-cpu/repack.h
         ggml-cpu/ggml-cpu-hbm.cpp
         ggml-cpu/ggml-cpu-hbm.h
         ggml-cpu/ggml-cpu-quants.c
